# typofix in class TranslateTable ! framework/functions.php set function create_access_file to deprecated ! set ModLanguage.php to deprecated ! add methode escapeString  to class WbDatabase ! update upgrade-script.php ! add getter property LastInsertId to class WbDatabase

+++ wb/modules/news/upgrade.php
@@ -1,18 +1,38 @@
 <?php
 /**
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  *
- * @category        modules
- * @package         news
- * @subpackage      upgrade
- * @copyright       2009-2012, WebsiteBaker Org. e.V.
- * @link			http://www.websitebaker2.org/
- * @license         http://www.gnu.org/licenses/gpl.html
- * @platform        WebsiteBaker 2.8.x
- * @requirements    PHP 5.2.2 and higher
- * @version         $Id$
- * @filesource		$HeadURL$
- * @lastmodified    $Date$
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
  *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+/**
+ * upgrade.php
+ * 
+ * @category     Module
+ * @package      Module_news
+ * @subpackage   upgrade
+ * @author       Dietmar Wöllbrink <dietmar.woellbrink@websitebaker.org>
+ * @author       Werner v.d.Decken <wkl@isteam.de>
+ * @copyright    Werner v.d.Decken <wkl@isteam.de>
+ * @license      http://www.gnu.org/licenses/gpl.html   GPL License
+ * @version      0.0.1
+ * @revision     $Revision$
+ * @link         $HeadURL$
+ * @lastmodified $Date$
+ * @since        File available since 17.01.2013
+ * @description  xyz
+ * 
  */
 
 /* -------------------------------------------------------- */
@@ -26,99 +46,83 @@ if(!function_exists('mod_news_Upgrade'))
 	function mod_news_Upgrade()
 	{
 		global $database, $admin, $MESSAGE,$bDebugModus;
-    	$msg = array();
+		$msg = array();
 		$callingScript = $_SERVER["SCRIPT_NAME"];
 // check if upgrade startet by upgrade-script to echo a message
 		$tmp = 'upgrade-script.php';
 		$globalStarted = substr_compare($callingScript, $tmp,(0-strlen($tmp)),strlen($tmp)) === 0;
-        /**
-         * check database engine
-         */
-    	$aTable = array('mod_news_posts','mod_news_groups','mod_news_comments','mod_news_settings');
-         for($x=0; $x<sizeof($aTable);$x++) {
-        	if(($sOldType = $database->getTableEngine(TABLE_PREFIX.$aTable[$x]))) {
-        		if(('myisam' != strtolower($sOldType))) {
-        			if(!$database->query('ALTER TABLE `'.TABLE_PREFIX.$aTable[$x].'` Engine = \'MyISAM\' ')) {
-        				$msg[] = $database->get_error();
-        			} else{
-                        $msg[] = 'TABLE `'.TABLE_PREFIX.$aTable[$x].'` changed to Engine = \'MyISAM\'';
-        			}
-        		} else {
-                 $msg[] = 'TABLE `'.TABLE_PREFIX.$aTable[$x].'` has Engine = \'MyISAM\'';
-        		}
-        	} else {
-        		$msg[] = $database->get_error();
-        	}
-        }
-
+		/**
+		 * check database engine
+		 */
+		$aTable = array('mod_news_posts','mod_news_groups','mod_news_comments','mod_news_settings');
+		 for($x=0; $x<sizeof($aTable);$x++) {
+			if(($sOldType = $database->getTableEngine(TABLE_PREFIX.$aTable[$x]))) {
+				if(('myisam' != strtolower($sOldType))) {
+					if(!$database->query('ALTER TABLE `'.TABLE_PREFIX.$aTable[$x].'` Engine = \'MyISAM\' ')) {
+						$msg[] = $database->get_error();
+					} else{
+		                $msg[] = 'TABLE `'.TABLE_PREFIX.$aTable[$x].'` changed to Engine = \'MyISAM\'';
+					}
+				} else {
+							 $msg[] = 'TABLE `'.TABLE_PREFIX.$aTable[$x].'` has Engine = \'MyISAM\'';
+				}
+			} else {
+				$msg[] = $database->get_error();
+			}
+		}
 
 		$sPagesPath = WB_PATH.PAGES_DIRECTORY;
 		$sPostsPath = $sPagesPath.'/posts';
 	// create /posts/ - directory if not exists
-		if(!file_exists($sPostsPath)) {
-			if(is_writable($sPagesPath)) {
-				make_dir(WB_PATH.PAGES_DIRECTORY.'/posts/');
-			} else {
-				if(!$globalStarted){
-					$msg[] = ($MESSAGE['PAGES_CANNOT_CREATE_ACCESS_FILE']);
-//					$admin->print_error($MESSAGE['PAGES_CANNOT_CREATE_ACCESS_FILE']);
-					return $msg;
-				} else {
-					echo  ''.$MESSAGE['PAGES_CANNOT_CREATE_ACCESS_FILE'].'';
-					return $msg;
-				}
+		if(is_writable($sPagesPath)) {
+			if(!($bRetval = is_dir($sPostsPath))) {
+				$iOldUmask = umask(0) ;
+				// sanitize directory mode to 'o+rwx/g+x/u+x' and create path
+				$bRetval = mkdir($sPostsPath, (OCTAL_DIR_MODE |0711), true); 
+				umask($iOldUmask);
 			}
-            $msg[] = '<strong>Directory "'.PAGES_DIRECTORY.'/posts/" created</strong>';
-		} else {
-            $msg[] = '<strong>Directory "'.PAGES_DIRECTORY.'/posts/" already exists.</strong>';
+			if($bRetval) {
+				$msg[] = 'Directory "'.PAGES_DIRECTORY.'/posts/" already exists or created.';
+			}else {
+				$msg[] = ($MESSAGE['PAGES_CANNOT_CREATE_ACCESS_FILE']);
+			}
+		}else {
+				$msg[] = ($MESSAGE['PAGES_CANNOT_CREATE_ACCESS_FILE']);
 		}
 	// check if new fields must be added
 		$doImportDate = true;
 		if(!$database->field_exists(TABLE_PREFIX.'mod_news_posts', 'created_when')) {
 			if(!$database->field_add(TABLE_PREFIX.'mod_news_posts', 'created_when',
 			                        'INT NOT NULL DEFAULT \'0\' AFTER `commenting`')) {
-				if($globalStarted){
-					echo ''.$MESSAGE['RECORD_MODIFIED_FAILED'].'';
-					return $msg;
-				} else {
-//					$admin->print_error($MESSAGE['RECORD_MODIFIED_FAILED']);
 					$msg[] = $MESSAGE['RECORD_MODIFIED_FAILED'];
-					return $msg;
-				}
+			} else {
+				$msg[] = 'TABLE `'.TABLE_PREFIX.'mod_news_posts` Datafield `created_when` added.';
 			}
-
-            if($globalStarted) {
-                echo 'Datafield `'.TABLE_PREFIX.'mod_news_posts`.`created_when` added.<br />';
-            }
-		} else { $doImportDate = false; }
+		} else { 
+			$msg[] = 'TABLE `'.TABLE_PREFIX.'mod_news_posts` Datafield `created_when` already exists.';
+			$doImportDate = false; 
+		}
 
 		if(!$database->field_exists(TABLE_PREFIX.'mod_news_posts', 'created_by')) {
 			if(!$database->field_add(TABLE_PREFIX.'mod_news_posts', 'created_by',
 			                        'INT NOT NULL DEFAULT \'0\' AFTER `created_when`')) {
-				if($globalStarted){
-					echo ''.$MESSAGE['RECORD_MODIFIED_FAILED'].'';
-					return $msg;
-				}else {
-//					$admin->print_error($MESSAGE['RECORD_MODIFIED_FAILED']);
-					$msg[] = $MESSAGE['RECORD_MODIFIED_FAILED'];
-					return $msg;
-				}
+				$msg[] = $MESSAGE['RECORD_MODIFIED_FAILED'];
 			}
-            if($globalStarted) {
-                echo 'datafield `'.TABLE_PREFIX.'mod_news_posts`.`created_by` added.<br />';
-            }
+		} else { 
+			$msg[] = 'TABLE `'.TABLE_PREFIX.'mod_news_posts` Datafield `created_by` already exists.';
+			$doImportDate = false; 
 		}
-	// preset new fields `created_by` and `created_when` from existing values
+ 	// preset new fields `created_by` and `created_by` from existing values
 		if($doImportDate) {
 			$sql  = 'UPDATE `'.TABLE_PREFIX.'mod_news_posts` ';
 			$sql .= 'SET `created_by`=`posted_by`, `created_when`=`posted_when`';
 			$database->query($sql);
 		}
 
-    /**
-     * rebuild news post folder
-     */
-	$array = rebuildFolderProtectFile($sPostsPath);
+	/**
+	 * rebuild news post folder
+	 */
+//	$array = rebuildFolderProtectFile($sPostsPath);
 	// now iterate through all existing accessfiles,
 	// write its creation date into database
 		$oDir = new DirectoryIterator($sPostsPath);
@@ -137,17 +141,18 @@ if(!function_exists('mod_news_Upgrade'))
 					$sql  = 'UPDATE `'.TABLE_PREFIX.'mod_news_posts` ';
 					$sql .= 'SET `created_when`='.$fileinfo->getMTime().' ';
 					$sql .= 'WHERE `link`=\''.$link.'\'';
-					$database->query($sql);
+					if($database->query($sql)) {
+						// delete old access file
+						unlink($fileinfo->getPathname());
+						$count++;
+					}
 				}
-			// delete old access file
-				unlink($fileinfo->getPathname());
-				$count++;
 			}
 		}
 		unset($oDir);
 
 		if($count > 0) {
-			$msg[] = 'save date of creation from '.$count.' old accessfiles and delete these files.';
+			$msg[] = 'Save date of creation from '.$count.' old accessfiles and delete these files.';
 		}
 // ************************************************
 	// Check the validity of 'create-file-timestamp' and balance against 'posted-timestamp'
@@ -161,41 +166,6 @@ if(!function_exists('mod_news_Upgrade'))
 		$database->query($sql);
 // ************************************************
 
-	// rebuild all access-files
-		$count = 0;
-		$backSteps = preg_replace('@^'.preg_quote(WB_PATH).'@', '', $sPostsPath);
-		$backSteps = str_repeat( '../', substr_count($backSteps, '/'));
-		$sql  = 'SELECT `page_id`,`post_id`,`section_id`,`link` ';
-		$sql .= 'FROM `'.TABLE_PREFIX.'mod_news_posts`';
-		$sql .= 'WHERE `link` != \'\'';
-		if( ($resPosts = $database->query($sql)) )
-		{
-			while( $recPost = $resPosts->fetchRow() )
-			{
-				$file = $sPagesPath.$recPost['link'].PAGE_EXTENSION;
-				$content =
-					'<?php'."\n".
-					'// *** This file is generated by WebsiteBaker Ver.'.VERSION."\n".
-					'// *** Creation date: '.date('c')."\n".
-					'// *** Do not modify this file manually'."\n".
-					'// *** WB will rebuild this file from time to time!!'."\n".
-					'// *************************************************'."\n".
-					"\t".'$page_id    = '.$recPost['page_id'].';'."\n".
-					"\t".'$section_id = '.$recPost['section_id'].';'."\n".
-					"\t".'$post_id    = '.$recPost['post_id'].';'."\n".
-					"\t".'$post_section = '.$recPost['section_id'].';'."\n".
-					"\t".'require(\''.$backSteps.'index.php\');'."\n".
-					'// *************************************************'."\n";
-				if( file_put_contents($file, $content) !== false ) {
-				// Chmod the file
-					change_mode($file);
-				}else {
-					$msg[] = ($MESSAGE['PAGES_CANNOT_CREATE_ACCESS_FILE']);
-				}
-				$count++;
-			}
-            $msg[] = '<strong>Number of new created access files: '.$count.'</strong>';
-		}
         // only for upgrade-script
         if($globalStarted) {
             if($bDebugModus) {
@@ -203,15 +173,24 @@ if(!function_exists('mod_news_Upgrade'))
                     echo '<strong>'.$title.'</strong><br />';
                 }
             }
-        }
-        return $msg;
+        } 
+        return ( ($globalStarted==true ) ? $globalStarted : $msg);
 	}
 }
 // end mod_news_Upgrade
 
 // ------------------------------------
-
-$msg = mod_news_Upgrade();
-
-
+// only show if manuell upgrade
+if( is_array($msg = mod_news_Upgrade()) ) {
+	$sModulReorg = 'm_news_Reorg';
+	if(class_exists($sModulReorg)) {
+		$oReorg = new $sModulReorg();
+		$msg = array_merge($msg, $oReorg->execute() ); // show details
+//		$msg = array_merge($msg,(new $sModulReorg())->execute()); // show details
+	}
+    foreach($msg as $title) {
+        echo '<strong>'.$title.'</strong><br />';
+    }
+	echo '<strong>News upgrade finished </strong><br /><br>';
+}
 /* **** END UPGRADE ********************************************************* */
